Clean: add, rename several files, move fncs
