+++ src/turk_algo.c
@@ -6,269 +6,11 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 11:35:50 by mbartos           #+#    #+#             */
-/*   Updated: 2024/01/13 18:20:18 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/01/13 18:39:32 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
-
-void	numof_r_rr_moves(t_node *stck, int num_to_move, int *rx, int *rrx)
-{
-	int	numof_rx;
-	int	numof_rrx;
-	int	before_num_to_find;
-
-	before_num_to_find = 1;
-	numof_rx = 0;
-	numof_rrx = 0;
-	while (stck != NULL)
-	{
-		if (stck->number == num_to_move)
-			before_num_to_find = 0;
-		else if (before_num_to_find == 1)
-			numof_rx++;
-		else
-			numof_rrx++;
-		stck = stck->next;
-	}
-	if (numof_rx > 0 || numof_rrx > 0)
-		numof_rrx++;
-	*rx = numof_rx;
-	*rrx = numof_rrx;
-}
-
-int	rx_operations(t_poss *possibilities)
-{
-	if (possibilities->ra >= possibilities->rb)
-		return (possibilities->ra);
-	else
-		return (possibilities->rb);
-}
-
-int	rrx_operations(t_poss *possibilities)
-{
-	if (possibilities->rra >= possibilities->rrb)
-		return (possibilities->rra);
-	else
-		return (possibilities->rrb);
-}
-
-void	ra_rb_is_lowest(t_poss *possblts, int ra_rb_oper)
-{
-	if (possblts->ra >= possblts->rb)
-	{
-		possblts->operations = ra_rb_oper;
-		possblts->rr = possblts->rb;
-		possblts->ra = possblts->ra - possblts->rb;
-		possblts->rb = 0;
-		possblts->rra = 0;
-		possblts->rrb = 0;
-	}
-	else
-	{
-		possblts->operations = ra_rb_oper;
-		possblts->rr = possblts->ra;
-		possblts->rb = possblts->rb - possblts->ra;
-		possblts->ra = 0;
-		possblts->rra = 0;
-		possblts->rrb = 0;
-	}
-}
-
-void	rra_rrb_is_lowest(t_poss *possblts, int rra_rrb_oper)
-{
-	if (possblts->rra >= possblts->rrb)
-	{
-		possblts->operations = rra_rrb_oper;
-		possblts->ra = 0;
-		possblts->rb = 0;
-		possblts->rrr = possblts->rrb;
-		possblts->rra = possblts->rra - possblts->rrb;
-		possblts->rrb = 0;
-	}
-	else
-	{
-		possblts->operations = rra_rrb_oper;
-		possblts->ra = 0;
-		possblts->rb = 0;
-		possblts->rrr = possblts->rra;
-		possblts->rrb = possblts->rrb - possblts->rra;
-		possblts->rra = 0;
-	}
-}
-
-void	rr_rrr_optimaze(t_poss *possblts)
-{
-	int	rra_rb;
-	int	ra_rrb;
-	int	ra_rb;
-	int	rra_rrb;
-
-	ra_rb = rx_operations(possblts);
-	ra_rrb = possblts->ra + possblts->rrb;
-	rra_rb = possblts->rra + possblts->rb;
-	rra_rrb = rrx_operations(possblts);
-	if (ra_rb <= ra_rrb && ra_rb <= rra_rb && ra_rb <= rra_rrb)
-		ra_rb_is_lowest(possblts, ra_rb);
-	else if (rra_rrb <= ra_rb && rra_rrb <= ra_rrb && rra_rrb <= rra_rb)
-		rra_rrb_is_lowest(possblts, rra_rrb);
-	else if (ra_rrb <= ra_rb && ra_rrb <= rra_rb && ra_rrb <= rra_rrb)
-	{
-		possblts->operations = ra_rrb;
-		possblts->rra = 0;
-		possblts->rb = 0;
-	}
-	else if (rra_rb <= ra_rb && rra_rb <= ra_rrb && rra_rb <= rra_rrb)
-	{
-		possblts->operations = rra_rb;
-		possblts->ra = 0;
-		possblts->rrb = 0;
-	}
-}
-
-/**
- * @brief Calculates the number of operations needed for various movements.
- *
- * This function determines the number of operations required for both rotate 
- * (ra, rb) and * reverse rotate (rra, rrb) movements in the context of pushing
- * a number from stack_a to stack_b.
- *
- * @param possblts Pointer to the structure storing the number of operations
- * for each possibility.
- * @param stck_a Pointer to the top of stack_a (source stack).
- * @param stck_b Pointer to the top of stack_b (destination stack).
- * @param n_a Number to be pushed from stack_a to stack_b.
- * @param n_b Number in stack_b that affects the pushing strategy.
- */
-void	get_numof_oper(t_poss	*possblts, t_node *stck_a, t_node *stck_b)
-{
-	init_possblts(possblts);
-	numof_r_rr_moves(stck_a, possblts->num_a, &possblts->ra, &possblts->rra);
-	numof_r_rr_moves(stck_b, possblts->num_b, &possblts->rb, &possblts->rrb);
-	rr_rrr_optimaze(possblts);
-}
-
-int	lower_num(t_node *stck, int desired_num)
-{
-	int	closest_lower;
-
-	closest_lower = INT_MIN;
-	while (stck != NULL)
-	{
-		if (desired_num > stck->number && stck->number > closest_lower)
-			closest_lower = stck->number;
-		stck = stck->next;
-	}
-	return (closest_lower);
-}
-
-int	higher_num(t_node *stck, int desired_num)
-{
-	int	closest_higher;
-
-	closest_higher = INT_MAX;
-	while (stck != NULL)
-	{
-		if (desired_num < stck->number && stck->number < closest_higher)
-			closest_higher = stck->number;
-		stck = stck->next;
-	}
-	return (closest_higher);
-}
-
-void	do_operations(t_poss *lowest_possblts, t_node **stck_a, t_node **stck_b)
-{
-	int	i;
-
-	i = 0;
-	while (i++ < lowest_possblts->ra)
-		ra(stck_a);
-	i = 0;
-	while (i++ < lowest_possblts->rb)
-		rb(stck_b);
-	i = 0;
-	while (i++ < lowest_possblts->rr)
-		rr(stck_a, stck_b);
-	i = 0;
-	while (i++ < lowest_possblts->rra)
-		rra(stck_a);
-	i = 0;
-	while (i++ < lowest_possblts->rrb)
-		rrb(stck_b);
-	i = 0;
-	while (i++ < lowest_possblts->rrr)
-		rrr(stck_a, stck_b);
-}
-
-/**
- * @brief Pushes numbers from stack_b back to stack_a in an optimal manner.
- *
- * This function efficiently pushes numbers from stack_b back to stack_a by
- *  considering the optimal * sequence of reverse and rotate operations 
- * (ra, rra) to reach the right position for insertion.
- *
- * @param stck_a Pointer to the top of stack_a (destination stack).
- * @param stck_b Pointer to the top of stack_b (source stack).
- */
-void	push_stck_b_back_to_a(t_node **stck_a, t_node **stck_b)
-{
-	int		closest_higher;
-	t_poss	possblts;
-
-	init_possblts(&possblts);
-	while (*stck_b != NULL)
-	{
-		init_possblts(&possblts);
-		if ((*stck_b)->number > ft_stck_max(*stck_a))
-			closest_higher = ft_stck_min(*stck_a);
-		else
-			closest_higher = higher_num(*stck_a, (*stck_b)->number);
-		numof_r_rr_moves(*stck_a, closest_higher, &possblts.ra, &possblts.rra);
-		if (possblts.ra <= possblts.rra)
-			possblts.rra = 0;
-		else
-			possblts.ra = 0;
-		do_operations(&possblts, stck_a, stck_b);
-		pa(stck_a, stck_b);
-	}
-}
-
-/**
- * @brief Performs the last sorting step on stack_a.
- *
- * This function evaluates the optimal sequence of reverse and rotate operations
- *  (ra, rra) on stack_a * to achieve the final sorting state. It considers the 
- * number of moves needed for both operations * and executes the one with the 
- * minimum moves.
- *
- * @param stck_a Pointer to the top of stack_a (source stack).
- */
-void	last_sort_of_stck_a(t_node **stck_a)
-{
-	t_poss	possblts;
-
-	init_possblts(&possblts);
-	possblts.num_a = ft_stck_min(*stck_a);
-	numof_r_rr_moves(*stck_a, possblts.num_a, &possblts.ra, &possblts.rra);
-	if (possblts.ra <= possblts.rra)
-		possblts.rra = 0;
-	else
-		possblts.ra = 0;
-	do_operations(&possblts, stck_a, NULL);
-}
-
-void	possible_cpy(t_poss *dest, const t_poss *src)
-{
-	dest->num_a = src->num_a;
-	dest->num_b = src->num_b;
-	dest->operations = src->operations;
-	dest->ra = src->ra;
-	dest->rb = src->rb;
-	dest->rr = src->rr;
-	dest->rra = src->rra;
-	dest->rrb = src->rrb;
-	dest->rrr = src->rrr;
-}
 
 /**
  * @brief Pushes the number with the minimum operations needed from
@@ -315,7 +57,7 @@ void	push_one_num_from_a_to_b(t_node **stck_a, t_node **stck_b)
 	t_node	*stck_a_iter;
 
 	stck_a_iter = *stck_a;
-	init_possblts(&lowest_possblts);
+	possblts_init(&lowest_possblts);
 	lowest_possblts.operations = INT_MAX;
 	while (stck_a_iter != NULL)
 	{
@@ -327,9 +69,9 @@ void	push_one_num_from_a_to_b(t_node **stck_a, t_node **stck_b)
 			possblts.num_b = ft_stck_max(*stck_b);
 		else
 			possblts.num_b = lower_num(*stck_b, num);
-		get_numof_oper(&possblts, *stck_a, *stck_b);
+		get_numof_opers(&possblts, *stck_a, *stck_b);
 		if (possblts.operations < lowest_possblts.operations)
-			possible_cpy(&lowest_possblts, &possblts);
+			possblts_cpy(&lowest_possblts, &possblts);
 		stck_a_iter = stck_a_iter->next;
 	}
 	do_operations(&lowest_possblts, stck_a, stck_b);
@@ -354,6 +96,63 @@ void	pushing_nums_from_a_to_b(t_node **stck_a, t_node **stck_b)
 	}
 	while (ft_stcksize(*stck_a) > 3)
 		push_one_num_from_a_to_b(stck_a, stck_b);
+}
+
+/**
+ * @brief Pushes numbers from stack_b back to stack_a in an optimal manner.
+ *
+ * This function efficiently pushes numbers from stack_b back to stack_a by
+ *  considering the optimal * sequence of reverse and rotate operations 
+ * (ra, rra) to reach the right position for insertion.
+ *
+ * @param stck_a Pointer to the top of stack_a (destination stack).
+ * @param stck_b Pointer to the top of stack_b (source stack).
+ */
+void	push_stck_b_back_to_a(t_node **stck_a, t_node **stck_b)
+{
+	int		closest_higher;
+	t_poss	possblts;
+
+	possblts_init(&possblts);
+	while (*stck_b != NULL)
+	{
+		possblts_init(&possblts);
+		if ((*stck_b)->number > ft_stck_max(*stck_a))
+			closest_higher = ft_stck_min(*stck_a);
+		else
+			closest_higher = higher_num(*stck_a, (*stck_b)->number);
+		numof_r_rr_moves(*stck_a, closest_higher, &possblts.ra, &possblts.rra);
+		if (possblts.ra <= possblts.rra)
+			possblts.rra = 0;
+		else
+			possblts.ra = 0;
+		do_operations(&possblts, stck_a, stck_b);
+		pa(stck_a, stck_b);
+	}
+}
+
+/**
+ * @brief Performs the last sorting step on stack_a.
+ *
+ * This function evaluates the optimal sequence of reverse and rotate operations
+ *  (ra, rra) on stack_a * to achieve the final sorting state. It considers the 
+ * number of moves needed for both operations * and executes the one with the 
+ * minimum moves.
+ *
+ * @param stck_a Pointer to the top of stack_a (source stack).
+ */
+void	last_sort_of_stck_a(t_node **stck_a)
+{
+	t_poss	possblts;
+
+	possblts_init(&possblts);
+	possblts.num_a = ft_stck_min(*stck_a);
+	numof_r_rr_moves(*stck_a, possblts.num_a, &possblts.ra, &possblts.rra);
+	if (possblts.ra <= possblts.rra)
+		possblts.rra = 0;
+	else
+		possblts.ra = 0;
+	do_operations(&possblts, stck_a, NULL);
 }
 
 /**
